Refactor clear selection behaviour to a React hook
