+++ src/javascript/JContent/ContentRoute/ContentLayout/EmptyDropZone/EmptyDropZone.jsx
@@ -94,7 +94,6 @@ const EmptyDropZone = ({component: Component, isCanDrop, allowDrop, uploadType, 
 EmptyDropZone.propTypes = {
     component: PropTypes.string.isRequired,
     uploadType: PropTypes.string,
-    // eslint-disable-next-line react/boolean-prop-naming
     isCanDrop: PropTypes.bool,
     allowDrop: PropTypes.bool,
     selector: PropTypes.func
